pyproject.toml
[tool.poetry]
name = "kaggle_riiid"
version = "0.1.0"
description = ""
authors = ["hakubishin3 <shu@wantedly.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.4"
pandas-gbq = "^0.14.1"
google-cloud-storage = "^1.33.0"
scikit-learn = "^0.23.2"
numpy = "^1.19.4"
scipy = "^1.5.4"
matplotlib = "^3.3.3"
click = "^7.1.2"
torch = "^1.7.0"
transformers = "^3.5.1"
pyarrow = "^2.0.0"
lightgbm = "^3.1.0"
google-cloud-bigquery-storage = "0.8.0"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
